Lib

Documentation

class (Functor f, Functor g) => Push f g where

Minimal complete definition

part, pver

Methods

part :: f (g a) -> f (f (g a))

pver :: f (f (g a)) -> f (g (f a))

push :: f (g a) -> f (g (f a))

push = pver . part

class Pull f g where

Minimal complete definition

draw, bond

Methods

draw :: f (g (f a)) -> f (f (g a))

bond :: f (f (g a)) -> f (g a)

pull :: f (g (f a)) -> f (g a)

pull = bond . draw

newtype a :~ b infixr 1

Constructors

Ty

Fields

getTy :: a

data Cat a where

Constructors

Obj :: a -> Cat (a :~ Cat a)

Mor :: (a -> b) -> Cat ((a -> b) :~ (Cat a -> Cat b))

Comp :: Cat ((b -> c) :~ (Cat b -> Cat c)) -> Cat ((a -> b) :~ (Cat a -> Cat b)) -> Cat ((a -> c) :~ (Cat a -> Cat c))

Id :: Cat ((a -> a) :~ (Cat a -> Cat a))

fib0 :: (Num p, Num a, Eq a) => a -> p

fib1 :: (Ord a, Num p, Num a) => a -> p

fib2 :: Integer -> Integer

fib1 is a source transformation of fib0, each having exponential complexity in n when interpreted naively, but they're the standard definition of the fibonacci numbers.

Compare the timing of fib0 with fib2:

λ> fib0 $ [0..30]
[0,1,1,2,3,5,8,13,21,34,55,89,144,233,377,610,987,1597,2584,4181,6765,10946,17711,28657,46368,75025,121393,196418,317811,514229,832040]
(8.97 secs, 2,413,020,008 bytes)

λ> fib2 $ [0..30]
[0,1,1,2,3,5,8,13,21,34,55,89,144,233,377,610,987,1597,2584,4181,6765,10946,17711,28657,46368,75025,121393,196418,317811,514229,832040]
(0.01 secs, 628,816 bytes)

In fact, fib2 stays fast up to pretty large values:

λ> length $ fib2 $ [0..10^8]
100000001
(3.52 secs, 16,000,374,552 bytes)

λ> length $ fib2 $ [0..10^9]
1000000001
(35.85 secs, 160,000,375,376 bytes)

Suggesting that it's roughly O(n) as we'd expect.

fib3 :: (Num d, Integral a) => a -> d

collatz0 :: (Integral a1, Num a2) => a1 -> a2

One implementation of the collatz function

collatz1 :: (Integral a1, Num a2) => a1 -> a2

Another implementation of the collatz function, using case and mod instead of even

collatz2 :: (Integral a1, Num a2) => a1 -> a2

collatz1 using divMod n 4 to expand the number of cases covered in a single case statement

noNothing :: Tree (Maybe a) -> Tree (Maybe a)

Drop all subtrees whose root is Nothing

noNothing2 :: Tree (Tree (Maybe a)) -> Tree (Tree (Maybe a))

Another implementation of noNothing, this one specialized to precomposition with Tree

forestMap :: (Forest a -> Forest a) -> Tree a -> Tree a

Map over the top-level Forest in a Tree

filterTree :: (a -> Bool) -> a -> Tree a -> Tree a

Recursively filter a Tree by values

isPure :: Free f a -> Bool

Is the input Pure?

quote :: Push f g => (f (f (g a)) -> f (f (g b))) -> f (g a) -> f (g (f b))

f'(expr)

unquote :: Pull f g => (f (f (g a)) -> f (f (g b))) -> f (g (f a)) -> f (g b)

f`(expr)

someFunc :: IO ()